moved min and max for integer to numbers
