+++ prestans3/types/integer.py
@@ -9,8 +9,6 @@
     :license: Apache 2.0, see LICENSE for more details.
 """
 
-from prestans3.errors import ValidationException
-
 from . import Number
 
 
@@ -25,21 +23,3 @@ class Integer(int, Number):
 
     def __init__(self, value, base=10):
         int.__init__(value, base)
-
-
-def _min(instance, config):
-    if instance < config:
-        raise ValidationException(instance.__class__,
-                                  "{} property is {}, however the configured minimum value is {}".format(
-                                      instance.__class__, instance, config))
-
-
-def _max(instance, config):
-    if instance > config:
-        raise ValidationException(instance.__class__,
-                                  "{} property is {}, however the configured maximum value is {}".format(
-                                      instance.__class__, instance, config))
-
-
-Integer.register_property_rule(_min, name="min")
-Integer.register_property_rule(_max, name="max")
