+++ tests/types/test_integer.py
@@ -56,31 +56,5 @@ def test_from_value_raises_value_error_on_non_int_subclass():
         Integer.from_value(0.3)
 
 
-def test_min_property_rule_works():
-    class __Model(Model):
-        my_int = Integer.property(min=1)
-
-    model = __Model.mutable()
-    model.my_int = 1
-    model.validate()
-    model.my_int = 0
-    with pytest.raises(ValidationException) as ex:
-        model.validate()
-    assert "{} property is {}, however the configured minimum value is {}".format(
-        Integer, 0, 1) in str(ex)
-
-
-def test_max_property_rule_works():
-    class __Model(Model):
-        my_int = Integer.property(max=1)
-
-    model = __Model.mutable()
-    model.my_int = 1
-    model.validate()
-    model.my_int = 2
-    with pytest.raises(ValidationException) as ex:
-        model.validate()
-    assert "{} property is {}, however the configured maximum value is {}".format(
-        Integer, 2, 1) in str(ex)
